move parsing into seperate module
