+++ examples/database.rs
@@ -22,7 +22,6 @@ fn main() {
     let path = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
         .join("examples")
         .join("example.pkl");
-    let value = pkl_rs::value_from_config::<Config>(path);
-
+    let value = pkl_rs::from_config::<Config>(path);
     println!("{:?}", value);
 }
