+++ src/api/parser.rs
@@ -1,0 +1,264 @@
+use std::collections::BTreeMap;
+
+use crate::pkl::{
+    self, non_primitive, IPklValue, ObjectMember, PklMod, PklNonPrimitive, PklPrimitive, PklValue,
+};
+
+#[cfg(feature = "trace")]
+use tracing::trace;
+
+fn parse_member_inner(
+    type_id: u64,
+    slots: &mut std::slice::Iter<rmpv::Value>,
+) -> anyhow::Result<ObjectMember> {
+    let ident = slots
+        .next()
+        .map(|v| {
+            v.as_str()
+                .expect(&format!("expected str for ident, got {:?}", v))
+                .to_owned()
+        })
+        .unwrap();
+
+    let value = slots.next().expect("[parse_member_inner] expected value");
+
+    // nested object, map using the outer ident
+    if let rmpv::Value::Array(array) = value {
+        let pkl_value = eval_inner_bin_array(&array)?;
+        return Ok(ObjectMember(type_id, ident, pkl_value));
+    }
+
+    let primitive = parse_primitive_member(value)?;
+
+    Ok(ObjectMember(
+        type_id,
+        ident,
+        IPklValue::Primitive(primitive),
+    ))
+}
+
+fn parse_non_prim_member(type_id: u64, slots: &[rmpv::Value]) -> anyhow::Result<PklNonPrimitive> {
+    match type_id {
+        non_primitive::code::TYPED_DYNAMIC => {
+            let dyn_ident = slots[0].as_str().expect("expected fully qualified name");
+
+            let module_uri = slots[1].as_str().expect("expected module uri");
+
+            let members = slots[2].as_array().expect(&format!(
+                "expected array of abstract member objects, got: {:?}",
+                slots[2]
+            ));
+
+            let members = members
+                .iter()
+                .map(|m| parse_pkl_obj_member(&m.as_array().unwrap()))
+                .collect::<anyhow::Result<Vec<ObjectMember>>>()?;
+
+            return Ok(PklNonPrimitive::TypedDynamic(
+                type_id,
+                dyn_ident.to_owned(),
+                module_uri.to_owned(),
+                members,
+            ));
+        }
+        non_primitive::code::SET => {
+            let values = &slots[0];
+            let values = values.as_array().unwrap().to_vec();
+            let values = values
+                .iter()
+                .map(|v| parse_primitive_member(v))
+                .collect::<anyhow::Result<Vec<PklPrimitive>>>()?;
+            return Ok(PklNonPrimitive::Set(type_id, values));
+        }
+        non_primitive::code::MAPPING | non_primitive::code::MAP => {
+            let values = &slots[0];
+            let mut mapping: BTreeMap<String, PklValue> = BTreeMap::new();
+            let values = values.as_map().unwrap();
+            for (k, v) in values.iter() {
+                let key = k.as_str().expect("expected key for mapping");
+                if let Some(array) = v.as_array() {
+                    // parse the inner object
+                    if let IPklValue::NonPrimitive(PklNonPrimitive::TypedDynamic(
+                        _,
+                        _,
+                        _,
+                        members,
+                    )) = eval_inner_bin_array(array)?
+                    {
+                        let mut fields = BTreeMap::new();
+                        for member in members {
+                            let (ident, value) = member.to_pkl_value()?;
+                            fields.insert(ident, value);
+                        }
+
+                        mapping.insert(key.to_string(), PklValue::Map(fields));
+                    }
+                } else {
+                    mapping.insert(key.to_string(), parse_primitive_member(v)?.into());
+                }
+            }
+            return Ok(PklNonPrimitive::Mapping(type_id, PklValue::Map(mapping)));
+        }
+
+        non_primitive::code::LIST | non_primitive::code::LISTING => {
+            let values = &slots[0];
+            let values = values
+                .as_array()
+                .expect(&format!("Expected array, got {:?}", values))
+                .to_vec();
+            let values = values
+                .iter()
+                .map(|v| parse_primitive_member(v))
+                .collect::<anyhow::Result<Vec<PklPrimitive>>>()?;
+
+            return Ok(PklNonPrimitive::List(type_id, values));
+        }
+        non_primitive::code::DURATION
+        | non_primitive::code::DATA_SIZE
+        | non_primitive::code::PAIR
+        | non_primitive::code::INT_SEQ
+        | non_primitive::code::REGEX
+        | non_primitive::code::TYPE_ALIAS => {
+            todo!("type {} cannot be rendered as json", type_id);
+        }
+        _ => {
+            todo!("parse other non-primitive types. type_id: {}", type_id);
+        }
+    }
+}
+
+fn parse_primitive_member(value: &rmpv::Value) -> anyhow::Result<PklPrimitive> {
+    match value {
+        rmpv::Value::String(s) => {
+            let Some(s) = s.as_str() else {
+                return Err(anyhow::anyhow!("expected valid UTF-8 string, got {:?}", s));
+            };
+            return Ok(PklPrimitive::String(s.to_string()));
+        }
+        rmpv::Value::Boolean(b) => Ok(PklPrimitive::Bool(b.to_owned())),
+        rmpv::Value::Nil => Ok(PklPrimitive::Null),
+        rmpv::Value::Integer(n) => {
+            if n.is_i64() {
+                Ok(PklPrimitive::Int(pkl::Integer::Neg(n.as_i64().unwrap())))
+            } else if n.is_u64() {
+                Ok(PklPrimitive::Int(pkl::Integer::Pos(n.as_u64().unwrap())))
+            } else if n.as_f64().is_some() {
+                Ok(PklPrimitive::Float(n.as_f64().unwrap()))
+            } else {
+                return Err(anyhow::anyhow!("expected integer, got {:?}", n));
+            }
+        }
+        _ => {
+            todo!("parse other primitive types. value: {}", value);
+        }
+    }
+}
+
+fn eval_inner_bin_array(slots: &[rmpv::Value]) -> anyhow::Result<IPklValue> {
+    let type_id = slots
+        // .next()
+        [0]
+    .as_u64()
+    // .map(|v| v.as_u64().expect(&format!("expected type id, got {:?}", v)))
+    .expect("missing type id");
+
+    if type_id == non_primitive::code::OBJECT_MEMBER {
+        // next slot is the ident, we don't need rn bc it's in the object from the outer scope that called this function
+        let value = &slots[2];
+        let primitive = parse_primitive_member(value)?;
+        return Ok(IPklValue::Primitive(primitive));
+    }
+
+    let non_prim = parse_non_prim_member(type_id, &slots[1..])?;
+    Ok(IPklValue::NonPrimitive(non_prim))
+}
+
+fn parse_pkl_obj_member(data: &[rmpv::Value]) -> anyhow::Result<ObjectMember> {
+    let mut slots = data.iter();
+
+    let type_id = slots
+        .next()
+        .map(|v| v.as_u64().expect(&format!("expected type id, got {:?}", v)))
+        .expect("missing type id");
+
+    match type_id {
+        non_primitive::code::OBJECT_MEMBER => {
+            return parse_member_inner(type_id, &mut slots);
+        }
+        non_primitive::code::DYNAMIC_LISTING => {
+            return parse_dynamic_list_inner(type_id, &mut slots);
+        }
+        _ => {
+            todo!("type_id is not OBJECT_MEMBER, or DYNAMIC_LISTING. implement parse other non-primitive types. type_id: {}\n", type_id);
+        }
+    }
+}
+
+/// this function is used to parse dynmically typed listings
+
+/// i.e:
+/// ```
+/// birds = new {
+///  "Pigeon"
+///  "Hawk"
+///  "Penguin"
+///  }
+/// ```
+/// the dynamically typed listings have a different structure than the typed listings
+///
+/// TODO: there has to be a bug here
+fn parse_dynamic_list_inner(
+    type_id: u64,
+    slots: &mut std::slice::Iter<rmpv::Value>,
+) -> anyhow::Result<ObjectMember> {
+    #[cfg(feature = "trace")]
+    trace!("parse_dynamic_list_inner: type_id: {}", type_id);
+    if type_id != non_primitive::code::DYNAMIC_LISTING {
+        todo!(
+            "expected DYNAMIC_LISTING ( type_id: {}), got: {}",
+            non_primitive::code::DYNAMIC_LISTING,
+            type_id
+        );
+    }
+
+    let index = slots
+        .next()
+        .map(|v| v.as_u64().expect("expected index"))
+        .unwrap();
+
+    let value = slots.next().expect("[parse_member_inner] expected value");
+
+    // nested object, map using the outer ident
+    if let rmpv::Value::Array(array) = value {
+        let pkl_value = eval_inner_bin_array(&array)?;
+        return Ok(ObjectMember(type_id, index.to_string(), pkl_value));
+    }
+
+    let primitive = parse_primitive_member(value)?;
+
+    Ok(ObjectMember(
+        type_id,
+        index.to_string(),
+        IPklValue::Primitive(primitive),
+    ))
+}
+
+pub fn pkl_eval_module(decoded: &rmpv::Value) -> anyhow::Result<PklMod> {
+    let root = decoded.as_array().unwrap();
+    let module_name = root.get(1).expect("expected root level module name");
+    let module_uri = root.get(2).expect("expected root level module uri");
+    let children = root.last().expect("expected children");
+
+    let pkl_module = children.as_array().unwrap();
+
+    let members = pkl_module
+        .iter()
+        .map(|f| parse_pkl_obj_member(f.as_array().unwrap()))
+        .collect::<anyhow::Result<Vec<ObjectMember>>>()?;
+
+    Ok(PklMod {
+        _module_name: module_name.as_str().unwrap().to_string(),
+        _module_uri: module_uri.as_str().unwrap().to_string(),
+        members,
+    })
+}
